requirements.txt
Django==1.7.5
Pillow==2.7.0
coverage==3.7.1
django-appconf==1.0.1
django-compressor==1.4
django-extensions==1.5.0
django-extra-views==0.6.5
django-geojson==2.6.0
django-leaflet==0.15.2
django-registration-redux==1.1
factory-boy==2.4.1
geopy==1.9.1
psycopg2==2.6
requests==2.5.3
six==1.5.2
wsgiref==0.1.2
